enhance: the client request timeout with cpu idle time, now replaced with semaphore to communicate.

+++ rpc/src/main/java/io/github/rezeros/core/client/RpcReferenceWrapper.java
@@ -46,7 +46,7 @@ public class RpcReferenceWrapper<T> {
         this.attatchments = attatchments;
     }
 
-    public void setTimeOut(int timeOut) {
+    public void setTimeOut(long timeOut) {
         attatchments.put("timeOut", timeOut);
     }
